+++ R/as_nombre_comuna.R
@@ -7,17 +7,16 @@
 #' @returns Vector con nombres de comuna
 #' @export
 as_nombre_comuna <- function(codigos_comunas) {
-
-  if (!class(codigos_comunas) %in% c("numeric", "integer")) {
-   cli::cli_abort("códigos comunales deben venir en formato numérico")
+  if (!is.numeric(codigos_comunas)) {
+    cli::cli_abort("Códigos comunales deben estar en formato numérico")
   }
 
   nombres_comunas <- territorial::territorios$nombre_comuna
 
-  nombres_encontrados <- territorial::territorios$nombre_comuna[match(codigos_comunas, territorial::territorios$codigo_comuna)]
+  nombres_encontrados <- territorial::territorios$nombre_comuna[match(
+    codigos_comunas,
+    territorial::territorios$codigo_comuna
+  )]
 
   return(nombres_encontrados)
 }
-
-
-